tools/book_service/booksmcp/requirements.txt
# MCP Server - FastMCP for streamable HTTP transport
fastmcp>=0.5.0

# Database (from parent booksdb module)
pymysql>=1.1.0
numpy>=1.24.0
